scanner: Only initialise drive service if identity manager is defined
In theory, this should not happen in most cases:

- The Scanner PKSF depends on the identity manager PKSF.
- The Scanner PKSF only builds the service in regular profiles, and
  redirects the service in regular OTR profiles.
  The identity manager PKSF builds the service in regular profiles, but
  does *not* build the service in regular OTR profiles.
  This is fine - in a regular OTR profile, the Scanner PKSF would
  construct the service using the regular profile, which has an
  associated identity manager service.
- `IdentityManagerFactory::BuildServiceInstanceFor` never returns
  nullptr.

However, it could be possible that `GetForProfile` is called during
service shutdown, or some time when the identity manager service is
null. Catch this case to prevent a UAF.

This will make getting the drive service return nullptr, which is
gracefully handled by the action handler which uses it.

Bug: b:367872031, b:367870452
Change-Id: I6a6a21003986c4a63da8cd70242d2428fb7c8da0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5951590
Reviewed-by: Michelle Chen <michellegc@google.com>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1372538}