[ios] Simplify EnterpriseAppAgent
Convert it to a SceneObservingAppAgent and introduce helper methods
to make the code easier to read.

The code is app-specific, so update the description of the class to
make it clear that is wait for the device enterprise policy, not any
profile related policies.

Fixed: 364234793
Low-Coverage-Reason: Improving untested code.
Change-Id: I3e73255c5d3e75ba0f4fc295ec935ca74270c6e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5953329
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372576}