Fix autofill availability flag for kRetrievePredictionImprovements.
This CL fixes the following case: `AutofillSuggestionAvailability` was
not set when the popup prompting the user to use prediction
improvements suggestions is shown. Without this attribute set, there
is no clue that autofill is available in any form. With it, screen
reader users hear something like "... with autofill menu. Use up/down
keys..." when a field is focused.

Bug: 372447295
Change-Id: I387e4499c2a5867df39596a3f30e33bca1ffdc3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5941024
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Dmitry Vykochko <vykochko@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1372591}