[iOS][PageInfo] Egtests wait for url to be propagated to the history DB
This is a follow up to https://crrev.com/c/5947128. That CL doesn't seem
to have fix all the flakiness. I believe the rest of the flakiness is
due to not waiting for the url to propagate to the history DB. As such,
sometimes the newly added history is not available when the PageInfo UI
is start up.

This CL also takes the opportunity to remove adding a hint to the ATP
service on tests testing the last visited functionality.

Fixed: 374063042
Change-Id: I6fcc05d834dcc7a23c410075123dbdb9d12930e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5953390
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Commit-Queue: Filipa Senra <fsenra@google.com>
Cr-Commit-Position: refs/heads/main@{#1372599}