[sync] Fix check failure in chrome://sync-internals while configuring
The refactoring in https://crrev.com/c/5895510 overlooked the case that
DataTypeManager can be CONFIGURING (or a few other short-lived states)
when the user visits the node browser in chrome://sync-internals.

In such scenario, before this patch, GetAllNodesRequestBarrier could be
exercised with an empty set of types, leading to a CHECK failure in the
constructor.

Instead, this patch changes
DataTypeManagerImpl::GetAllNodesForDebugging() to generalize the cases
where an empty result must be returned without instantiating
GetAllNodesRequestBarrier. The condition to return early is identical
to what the CHECK requires which is simplest to reason about.

Change-Id: I1f0e08d7cc2fe7971bc37efeeaf954ef577eb517
Fixed: 374401600
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5956394
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1372610}