Clean up DisableForForbiddenUsers feature param
The feature parameter was introduced primarily to counter 403 errors
encountered during polling for unenrolled profiles in multi-user
setups.

We no longer have polling and we now have ways of checking whether a
given profile is enrolled in the Dogfood. As a result, we can remove
the feature parameter.

Bug: 322279583
Change-Id: Ia9047104ab4205924ef6ab56dc8f0eb34390844b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5952733
Reviewed-by: Norge Vizcay <vizcay@google.com>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1372613}