[iOS][Account Menu]Remove add account button
This CL is a temporary fix for crbug.com/374960362. It removes the Add
Account button in the "Account menu > Manage accounts on this device"
that causes a CHECK failure.

It should be reverted and properly fixed. This CL is mostly intended
for the purpose of cherry-picking it to M131.

fundamentally changed and was not tested previously

Low-Coverage-Reason: TRIVIAL_CHANGE The not-covered code is not
Bug: 374960362
Change-Id: Idd1c4ef6fe8976d357876c20c92cffa43284e10e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5953729
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372629}