Remove the feature flag check from FieldClassificationModelExecutor
This is needed to make the handler code reusable by PasswordManager.
The CHECK was excessive and happening in a unexpected place, since all model-related classes should not even be instantiated when the
corresponding feature flags are off.

Also, the model should never be loaded  when the feature is off, so
DCHECK at the prediction stage was already too late.

Bug: 371933424
Change-Id: I038f2198ef05675a37ce1bd180e92d24011735f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5952509
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1372642}