[ios] Fix StartSurfaceSceneAgent to async loading of Profile
When Profile are loaded asynchronously, it is possible for the
StartSurfaceSceneAgent to be created while the SceneState is
already in the foreground.

Call the method -sceneState:transitionedToActivationLevel: when
the SceneState is connected to ensure the StartSurfaceSceneAgent
correctly supports this case.

Bug: 333863468
Change-Id: I8eb8923a82ac3006e31d4f82494759a6c461d4cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5952932
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372653}