privacy-hub: Revert ArcGoogleLocationServicesEnabled policy description
New GoogleLocationServicesEnabled policy is still under development and
not published in the Admin Panel. This new policy deprecates the
ArcGoogleLocationServicesEnabled policy, but because it's not yet
published the documentation updates should not be released yet.

Reverting the updated documentation to the old version (c7a298a). This
should propagate to the Admin Panel and chromeenterprise.google/policies
surface. We'll undo this revert when GoogleLocationServicesEnabled is
rolled out in the Admin Panel (b/330216928).

Bug: b:372870090
Change-Id: Idbce6ba28b891f5dccfa7588483d06e143e23eb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5947444
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Zauri Meshveliani <zauri@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372677}