[ios] Do not remove old snapshot when taking a snapshot fails
Previously, the old snapshot image was deleted when taking a new image
on the same tab fails.
It caused the issue that we lost images on the tab grid.

Now, this CL updates `updateSnapshotStorage()` and
`-updateSnapshotStorageWithImage:` not to delete the old image when
snapshotting fails.
Callers should explicitly remove a snapshot.

Bug: 326359709
Change-Id: I6d10b77185245abdd57a590391e1012c3e493cd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5955696
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372689}