Align Omnibox size and font height across CTA and SA.
ChromeTabbedActivity (specifically: ToolbarPhone) computes the size of
the LocationBar visuals at runtime to offer smooth transition between
the focused and unfocused state, which is particularly important when
transitioning between NTP fakebox and active Omnibox. SearchActivity
relies on the same layout, however the transformations are not
applied, producing a smaller font and general misalignment between
the two activities.

This change ports LocationBar background and sizing from ToolbarPhone
to SearchActivity.

Note that the change does not align positioning yet - I will try to
address that in a separate change.

Related code:
https://crsrc.org/c/chrome/browser/ui/android/toolbar/java/src/org/chromium/chrome/browser/toolbar/top/ToolbarPhone.java;l=907-925

Screens:
- before: https://screenshot.googleplex.com/9CrBqk4eZ8QHgst.png
- after: https://screenshot.googleplex.com/5JmvFmLik7jVG8G.png

Bug: b:370570081
Change-Id: I648d4553ea6b7cb2bb8b6d88e973c32195623e25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5955571
Reviewed-by: Gang Wu <gangwu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372704}