Allow DISABLE_ for FUZZ_TESTs.
FUZZ_TESTs are a new type of test which can be inserted alongside
existing unit tests in order to cheaply increase fuzzing coverage.
They run on ClusterFuzz in full fuzzing mode, but they also run
briefly on our unit tests runners.

Like functional unit tests, they may sometimes be unreliable or
flakey, and we might want gardeners to be able to disable them.
The standard convention for a gardener to disable a test is to
rename it to be prefixed with DISABLED_. (Sometimes this is done
using a MAYBE_ macro so that it can be conditionally disabled
on some platforms or build configurations).

This convention doesn't quite work for FUZZ_TESTs, since (prior
to this change) the name of the test is given in three places -
twice in the C++ and once in the BUILD.gn.

This allows gardeners to disable FUZZ_TESTs with just two code
changes, instead of three: there is no longer a need to change
the BUILD.gn file.

When such a DISABLED_ test is encountered:
* it will be skipped over by our usual unit test runners, just
  like unit tests
* ClusterFuzz will be unable to run the test because it will
  encounter a message saying "No FUZZ_TEST matches the name X"

Other options were considered to solve this problem:
1) Require gardeners to disable FUZZ_TESTs in three places
2) Require gardeners to use an alternative way to disable
   FUZZ_TESTs, e.g. a new DISABLED_FUZZ_TEST macro
3) Complex template or preprocessor magic to allow FuzzTest to
   identify a function name without a DISABLED_ prefix; rejected
   because of the risk that we wouldn't spot genuine function
   name mismatches
4) Change gardenering entirely so that even unit tests are disabled
   a different way that doesn't involve tinkering with the name

It is not ideal that such DISABLED_ tests won't run on ClusterFuzz.
Ideally, we'd have them continue to run. This could be achieved
by altering the wrapper executable to check for DISABLED_
test names using --list_fuzz_tests, but at the moment we
don't consider this extra complexity is merited.

Change-Id: I323ca8d257b41c51a060c2c1fb6b52d082da2ea7
Cq-Include-Trybots: luci.chromium.try:linux-centipede-asan-rel,win-libfuzzer-asan-rel,linux-libfuzzer-asan-rel
Bug: 40286211
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5954268
Reviewed-by: Titouan Rigoudy <titouan@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372714}