Add TRACE events for setting/clearing touchscreen gesture target.
Debugging for the associated issue shows that scroll bubbling is failing
due to RWHIER::BubbleScrollEvent() thinking that a touchscreen gesture
event is currently in progress. Since the users are attempting touchpad
scrolls when the issue occurs, it's important to determine what is
setting the touchscreen target in RWHIER.

This CL adds TRACE events to detect the gesture type and target value
both when the touchscreen_gesture_target is set and when it is cleared.

Bug: 346629231
Change-Id: Ie384336237f2c810df5d05bb4bfc2887442da97c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5953464
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372775}