[BCIV] Add OffsetTag to scroll toolbar hairline without extra frames
The toolbar hairline should instantly disappear/appear when the toolbar is scrolled off/on screen. This is done by hiding/showing the slim::Layers for the composited toolbar, but a compositor frame is sent every time the visibility changes. We don't want these compositor frames to be sent, because it causes jank and increases latency.

In the current state of BCIV, we use one OffsetTag to move everything, meaning that the toolbar and viewport must move together by the same amount. The hairline is below the toolbar and z-indexed on top of the viewport, so if we don't send the compositor frames to hide the composited toolbar, the hairline will stay visible at the top of the screen. We fix this problem by scrolling only the toolbar layer a bit farther, so that the toolbar+hairline is moved completely off the screen without anything else moving incorrectly. This CL introduces another OffsetTag, so that the toolbar and viewport each have their own OffsetTag and can be scrolled with different offsets. This CL also removes the additional compositor frames for hiding/showing the hairline by making the composited toolbar always shown.

Bug: 373705114
Change-Id: If943c877b0c60cd2f15842d01a4f6bca080d4142
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5933670
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Cr-Commit-Position: refs/heads/main@{#1372804}