[Blink] Have RenderingContext2D create CanvasResourceProvider via host
Currently CanvasRenderingContext2D sometimes creates the
CanvasResourceProvider via its associated HTMLCanvasElement and
sometimes that object's Canvas2DLayerBridge. In practice these two flows
are the same, as HTMLCanvasElement will immediately call through to
Canvas2DLayerBridge to create the CanvasResourceProvider in the case
where its context is 2D (as it is by definition if
CanvasRenderingContext2D is being used).

This CL has CanvasRenderingContext2D always create the CRP via the
canvas element. This paves the way for a followup wherein we fold
Canvas2DLayerBridge's GetOrCreateResourceProvider() into
HTMLCanvasElement itself.

Note that in the future we have a goal to move CanvasResourceProvider
creation/ownership to the contexts, but that is for a later phase of
the overall canvas rearchitecture effort. Consolidating the current
distributed logic will facilitate that later phase.

Bug: 40280152
Change-Id: Id58b9571ccce748f5057feafdc671d5733f97c71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5946323
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372827}