[iOS] Create and present temporary welcome screens for passkeys
This CL adds the temporary enrollment and re-authentication welcome
screens to the SDS fetching flow. Those screens only contain a button
for now. They will eventually be more detailed and match the mocks, but
the goal here is just to add the step of integrating those screens to
the flow to enable end-to-end testing.

Video of the enrollment flow:
https://drive.google.com/file/d/1KPKnCC3zwZfHbaUGBYISV453evDRk6sN/view?usp=sharing&resourcekey=0-Zji1bZcw8ZWb5hoEEshEdw

Video of the re-authentication flow:
https://drive.google.com/file/d/1aNrQ0IDC3D1peqKDgQc6Bp6_w5ZQ9EEH/view?usp=sharing&resourcekey=0-zPlPTDgfmRatKHPkKSQn7A

Low-Coverage-Reason: TESTS_IN_SEPARATE_CL
Bug: 355042392
Change-Id: Ie5ac2b3bdac6fff2c43c2539aec43cbb8e0b7202
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5953845
Reviewed-by: Alexis Hétu <sugoi@chromium.org>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1372887}