Fix custom tab navigation observers for hidden tabs
Instead of missing or faking events for hidden tabs, initialize the
navigation observers for the hidden tab when it's created. If the speculated navigation is dropped, the observers are discarded (and page load metrics are only recorded when the speculation is actually used).

Pre-requisite to https://chromium-review.googlesource.com/c/chromium/src/+/5923757

Bug: 372710946
Change-Id: Ic30068ca8ce9eae31731f49d9d9e968fb242be48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5936843
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372959}