Fix most CrOS failures with is_ubsan_vptr
There are still two remaining, but this fixes most of it. In general,
the issue is that some parts of CrOS ended up with a pattern where an
unused class is left always null, but some casts into that unused class
are still there. is_ubsan_vptr adds references to those classes.

Broadly the fix is to push the disabled code up a layer and also exclude
those files. SecureChannelImpl::Factory I just removed entirely because
it never worked at all. The destructor was never being defined, but it
was fine because no one ever used it.

Bug: 373429173
Change-Id: I6c4bb913bdc28d66a04f5e3cc0fdd26ab77999c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5938468
Commit-Queue: David Benjamin <davidben@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1372977}