Move CreateOnDeviceSessionTask classes to a new file
This is a no-op refactor that moves the `CreateOnDeviceSessionTask` to
a new file.

We need this in the follow-up CL that refactors the creation flow,
but since we are changing thew implementation, it's better to create
this base CL so that it's easier for reviewer to understand what the
changes are.

Bug: 373519560
Change-Id: I46838bc73b88d94e9974fccfba20d13ca697c56b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5954150
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1373113}