updater: Hide `ParseTagImpl()`
This CL is motivated by wanting to resolve the tension between
`ParseResult` sitting on the cusp of being a complex struct (and thus
hostile to admitting a `raw_span` as a member) and the `raw_ptr`
rewriter wanting to rewrite it.

This CL is intended to be a no-op. It

*   hides `ParseTagImpl()` directly inside the implementation of
    `certificate_tag`. `ParseTagImpl()` is never used outside this
    translation unit, and so needs no declaration in headers.

*   tears down `ParseResult` as a type, replacing it with an
    `absl::variant` with equivalent types. This allows the tri-state
    logic to be expressed more directly, and obviates the need for
    `raw_span`, as it's immediately laundered away.

Fixed: 372311617
Change-Id: If2efbd72a485904f6d1a805c0502d56ae009160f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5953368
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1373116}